\name{chooseWEAddin}
\alias{chooseWEAddin}
\title{Choose Worked Example}
\usage{
chooseWEAddin()
}
\description{
This addin let's the user select a Worked Example (vignette) of the course "Landscape Genetic Data Analysis with R" and choose between different file types to open (html, Rmd, R).
}

