\name{dd.ecogen}
\docType{data}
\alias{dd.ecogen}
\title{Genetic data for the freshwater snail \emph{Drepanotrema depressissimum}}
\description{
  Data used in Lamy et al. (2012).
}
\usage{dd.ecogen}
\format{An 'ecogen' object with 1271 observations and the following slots:
\describe{
  \item{XY}{Spatial coordinates: \describe{
    \item{Longitude}{West (WGS84)}
    \item{Latitude}{North (WGS84)}
  }}
  \item{G}{Genotypic data for 10 microsatellite loci: \describe{
    \item{DDE16}{Genotypes at microsatellite locus DDE16 (two alleles, unseparated).}
    \item{DDE4}{Genotypes at microsatellite locus DDE4 (two alleles, unseparated).}
    \item{DDE9}{Genotypes at microsatellite locus DDE9 (two alleles, unseparated).}
    \item{DDE12}{Genotypes at microsatellite locus DDE12 (two alleles, unseparated).}
    \item{DDE22}{Genotypes at microsatellite locus DDE22 (two alleles, unseparated).}
    \item{DDE20}{Genotypes at microsatellite locus DDE20 (two alleles, unseparated).}
    \item{DDE06}{Genotypes at microsatellite locus DDE06 (two alleles, unseparated).}
    \item{DDE14}{Genotypes at microsatellite locus DDE14 (two alleles, unseparated).}
    \item{DDE23}{Genotypes at microsatellite locus DDE23 (two alleles, unseparated).}
    \item{DDE3}{Genotypes at microsatellite locus DDE3 (two alleles, unseparated).}
  }}
    \item{A}{\describe{Allele frequencies derived from slot @G}
  }
    \item{S}{Structures (note that some sites were sampled in multiple years): \describe{
    \item{SiteID}{Combination of SITE and YEAR}
    \item{SITE}{Site name (abbreviated)}
    \item{YEAR}{Year of sampling}
  }}
}}
\references{
Lamy, T., Pointier, J. P., Jarne, P. and David, P. (2012), Testing metapopulation dynamics using genetic, demographic and ecological data. Molecular Ecology, 21: 1394–1410. doi:10.1111/j.1365-294X.2012.05478.x}
\keyword{datasets}


