\name{installDGS}
\alias{installDGS}
\title{Install Packages}
\usage{
installDGS()
}
\description{
This function checks the RStudio and R versions and installs all suggested packages for 'LandGenCourse'. These packages are used by the vignettes in 'LandGenCourse'. Modified from code provided by Jeffrey Evans.
}

