\name{openCheatsheetAddin}
\alias{openCheatsheetAddin}
\title{Open Cheat Sheet}
\usage{
openCheatsheetAddin()
}
\description{
This addin let's the user select an R cheat sheet, download it to the 'downloads' folder in the active working directory, and open it in a PDF viewer. In the case of the list of R functions used in swirl course 'Landscape_Genetics_R_Course'), an editable DOC file is opened.
}

