\name{ralu.loci}
\docType{data}
\alias{ralu.loci}
\title{Subset of genetic data for Columbia spotted frog (\emph{Rana luteiventris})}
\description{
  Subset of data used in Murphy et al.(2010).
}
\usage{ralu.loci}
\format{A data frame with 181 observations and 10 variables:
\describe{
\item{SiteName}{Unique site name}
\item{Pop}{Short site name (for labels)}
\item{A}{Genotypes at microsatellite locus A. The two alleles are separated by ':'}
\item{B}{Genotypes at microsatellite locus B. The two alleles are separated by ':'}
\item{C}{Genotypes at microsatellite locus C. The two alleles are separated by ':'}
\item{D}{Genotypes at microsatellite locus D. The two alleles are separated by ':'}
\item{E}{Genotypes at microsatellite locus E. The two alleles are separated by ':'}
\item{F}{Genotypes at microsatellite locus F. The two alleles are separated by ':'}
\item{G}{Genotypes at microsatellite locus G. The two alleles are separated by ':'}
\item{H}{Genotypes at microsatellite locus H. The two alleles are separated by ':'}
}}
\references{
  Murphy M.A., R. Dezzani, D.S. Pilliod & A.S. Storfer (2010) Landscape genetics of high mountain frog metapopulations. Molecular Ecology 19(17):3634-3649}
\keyword{datasets}


