\name{watchVideoAddin}
\alias{watchVideoAddin}
\title{Watch a Course Video}
\usage{
watchVideoAddin()
}
\description{
This addin let's the user select a course video and open it from the web as a video (mp4), slides (pdf), or both. Slides are downloaded to the 'downloads' folder in the active working directory.
}

